requirements.txt
Django~=4.2
djangorestframework
djangorestframework-simplejwt
djoser
Faker
Markdown
mutagen
PyJWT
drf-spectacular